Add spark container

+++ docker/base/entrypoint.sh
@@ -1,5 +1,5 @@
-#!/bin/sh
-if [ -d $HOME/src ]; then
+#!/bin/bash
+if [[ -d $HOME/src && -f $HOME/src/setup.py ]]; then
     /opt/conda/bin/pip install --user --editable $HOME/src
 fi
 cd $HOME
